Add CodeByCommit method that returns code information grouped by commit.
This fixes a bug where commits without any files were not returned by Code call.
This is required to correctly get last processed commit.

When using old Code method if the last commit is a merge with no file changes,
it would be ignored and incremental run would fail with an error.